fix: clean up tooltip code
